pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "hackerone-mcp"
version = "1.0.0"
description = "HackerOne MCP Server - Model Context Protocol server for HackerOne integration"
authors = [
    {name = "Daniel Olson", email = "devbox4714@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "httpx>=0.28.1,<0.29.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "mcp>=1.9.0,<2.0.0",
    "PyYAML>=6.0.0,<7.0.0"
]

[tool.poetry]
name = "hackerone-mcp"
version = "1.0.0"
description = "HackerOne MCP Server - Model Context Protocol server for HackerOne integration"
authors = ["Daniel Olson <devbox4714@gmail.com>"]
readme = "README.md"
packages = [{include = "hackerone_mcp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.28.1"
python-dotenv = "^1.1.0"
mcp = "^1.9.0"
PyYAML = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"

[tool.poetry.scripts]
hackerone-mcp = "hackerone_mcp.main:run_server"
